refactor: move logout-related code from MainHeader to new LogoutButton component
