+++ frontend/src/App.jsx
@@ -1,3 +1,4 @@
+import { CircularProgress } from "@mui/material";
 import { QueryClientProvider } from "@tanstack/react-query";
 import { createBrowserRouter } from "react-router";
 import { RouterProvider } from "react-router/dom";
@@ -5,14 +6,12 @@ import ContextWrapper from "./pages/ContextWrapper";
 import GenericError from "./pages/GenericError.jsx";
 import HomePage from "./pages/Home";
 import LoginPage, { action as loginAction } from "./pages/Login";
-import LogoutError, { action as logoutAction } from "./pages/Logout";
+import PostCreationPage from "./pages/PostCreation.jsx";
 import PostDetailsPage from "./pages/PostDetails";
 import PrivacyPolicyPage from "./pages/PrivacyPolicy";
 import RootPage from "./pages/Root";
 import SignupPage, { action as signupAction } from "./pages/Signup";
 import { queryClient } from "./utils/http";
-import PostCreationPage from "./pages/PostCreation.jsx";
-import { CircularProgress } from "@mui/material";
 
 const router = createBrowserRouter([
   {
@@ -46,11 +45,6 @@ const router = createBrowserRouter([
         path: "/login",
         element: <LoginPage />,
         action: loginAction,
-      },
-      {
-        path: "/logout",
-        action: logoutAction,
-        errorElement: <LogoutError />,
       },
       {
         path: "/signup",
